"Better" code.

+++ io_utils.h
@@ -1,6 +1,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-size_t getSizeFromFileStream(FILE *stream);
+uint32_t getSizeFromFileStream(FILE *stream);
 char *readDataFromFileStream(FILE *stream);
